+++ settings.gradle
@@ -17,6 +17,7 @@ include ':app'
 include ':app-without-compose'
 include ':roborazzi-junit-rule'
 include ':roborazzi'
+include ':roborazzi-compose'
 include ':roborazzi-painter'
 
 includeBuild("include-build") {
